src/requirements.txt
functions-framework==3.*
openai
google-cloud-pubsub